index_options_bot/requirements.txt
dhanhq>=2.1.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b0
python-dotenv>=1.0.0
pytz>=2024.1